requirements.txt
aiogram==3.1.1
asyncpg==0.28.0
python-dotenv==1.0.0
solana==0.30.2
redis==5.0.0
aiohttp==3.8.5
pydantic==2.3.0
APScheduler==3.10.4
Pillow==10.0.0
matplotlib==3.7.2
numpy==1.25.2